+++ src/Symfony/Component/Validator/CHANGELOG.md
@@ -9,6 +9,7 @@ CHANGELOG
  * setting the `checkDNS` option of the `Url` constraint to `true` is deprecated in favor of
    the `Url::CHECK_DNS_TYPE_*` constants values and will throw an exception in Symfony 4.0
  * added min/max amount of pixels check to `Image` constraint via `minPixels` and `maxPixels`
+ * added a new "propertyPath" option to comparison constraints in order to get the value to compare from an array or object
 
 3.3.0
 -----